[Rename] : update directory structure
